requirements.txt
Flask==3.0.0
Werkzeug==3.0.1Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
google-auth-oauthlib==1.2.0
google-auth==2.28.0
python-dotenv==1.0.0